ata: ahci-imx: Covert to use GPIO descriptor
This converts the i.MX AHCI driver to use a GPIO descriptor
instead of parsing the device tree by itself.

This driver is quite obviously device tree only, and the
GPIO line is treated as optional, so let's keep it as
optional.

None of the device trees in the kernel use this GPIO
facility today, so it is hard to test.

Cc: Egor Starkov <egor.starkov@ge.com>
Cc: Richard Zhu <hongxing.zhu@nxp.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
---
ChangeLog v3->v4:
- Rebase on v5.2-rc1 and resend.
- If this keeps being missed I guess I'll just apply it to
  the GPIO tree.
ChangeLog v2->v3:
- Rebase on v5.1-rc1 and resend.
ChangeLog v1->v2:
- Drop the "clkreq had been initialized" message
- The message indicated several users on the same GPIO line,
  so set the GPIO as GPIOD_FLAGS_BIT_NONEXCLUSIVE so we can
  have more than one consumer of the same GPIO line