Input: tsc2007 - use GPIO descriptor
This switches the TSC2007 to use a GPIO descriptor to read
the pendown GPIO line.

As this will make the gpiolib start to respect polarity
inversion flags on the GPIO lines, drop the inversion when
reading the line with gpio_get_value(), fix two offenders
in the i.MX device trees, and also emphasize the importance
of marking the polarity right in the device tree bindings.

Cc: Denis Carikli <denis@eukrea.com>
Cc: Pengutronix Kernel Team <kernel@pengutronix.de>
Cc: NXP Linux Team <linux-imx@nxp.com>
Cc: H. Nikolaus Schaller <hns@goldelico.com>
Cc: devicetree@vger.kernel.org
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
---
ChangeLog v1->v2:
- Drop inversion on the GPIO descriptor value, rely on the
  gpiolib to handle polarity inversion.
- Comb through device trees, identify two offenders, fix
  them as part of the patch for a clean cut.
- Also fix the device tree bindings.