ARM: spitz: Use software nodes for the ADS7846 touchscreen
Convert the Spitz to use software nodes for SPI CS and the
ADS7846 touchscreen following the pattern used in the TI
platforms.

The Spitz is the only user of the wait_for_sync() callback in
the platform data, so define that this is a separate GPIO
in the device tree bindings (previous patch) and handle this
directly inside the ADS7846 driver instead.

Add some infrastructure to the PXA boardfile common code to
handle software nodes attached to the SPI bus instead of
platform data.

The SPI core will investigate the numner of GPIO handles to
figure out the number of chipselects used on the platform, so
no explicit encoding of the number of the number of chipselects
is needed.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>